refactor(router): clarify client boundaries
Keep router cache state direct and move lazy-slice bookkeeping beside Slice. Separate initial route and HMR hooks, and keep scroll and instant helpers in the history binding.
