+++ packages/waku/src/router/client-core-utils/caches.ts
@@ -24,106 +24,64 @@ export type PrefetchHandle = Pick<PrefetchEntry, 'promise' | 'onInvalidate'>;
 export const createRscParams = (query: string): URLSearchParams =>
   new URLSearchParams({ query });
 
-export const createCaches = () => {
-  const manager = createPrefetchManager();
-  const staticPathSet = new Set<string>();
-
-  const getPrefetchedElements = (route: RouteProps): Elements | undefined =>
-    manager.getElements(encodeRoutePath(route.path));
-
-  return {
-    prefetchRoute: (route: RouteProps, options?: PrefetchOptions): void => {
-      // skip is canReuseStaticRoute at the caller, which has this root's elements
-      const rscPath = encodeRoutePath(route.path);
-      manager.prefetch(
-        rscPath,
-        route.query,
-        (base, invalidate) =>
-          fetchRsc(rscPath, createRscParams(route.query), {
-            ...(base ? { unstable_base: base } : {}),
-            onBuildIdMismatch: () => {
-              invalidate();
-              manager.clear();
-            },
-          }),
-        options,
-      );
-    },
-    hasCachedShell: (
-      route: RouteProps,
-      currentElements: Record<string, unknown>,
-    ): boolean =>
-      canCommitInstantly(
-        getRouteSlotId(route.path),
-        currentElements,
-        getPrefetchedElements(route),
-      ),
-    getPrefetchedElements,
-    getPrefetch: (route: RouteProps): PrefetchHandle | undefined =>
-      manager.get(encodeRoutePath(route.path), route.query),
-    canReuseStaticRoute: (
-      route: RouteProps,
-      currentElements: Elements,
-    ): boolean =>
-      staticPathSet.has(route.path) &&
-      getRouteSlotId(route.path) in currentElements,
-    learnStaticFromElements: (elements: Record<string, unknown>): void => {
-      const route = getRouteFromElements(elements);
-      if (route && isStaticFromElements(elements)) {
-        staticPathSet.add(route.path);
-      }
-    },
-    clear: (): void => {
-      manager.clear();
-      staticPathSet.clear();
-    },
-  };
-};
-
-export type Caches = ReturnType<typeof createCaches>;
-
-const singleton = createCaches();
+const manager = createPrefetchManager();
+const staticPathSet = new Set<string>();
 
 export const prefetchRoute = (
   route: RouteProps,
   options?: PrefetchOptions,
-): void => singleton.prefetchRoute(route, options);
+): void => {
+  // skip is canReuseStaticRoute at the caller, which has this root's elements
+  const rscPath = encodeRoutePath(route.path);
+  manager.prefetch(
+    rscPath,
+    route.query,
+    (base, invalidate) =>
+      fetchRsc(rscPath, createRscParams(route.query), {
+        ...(base ? { unstable_base: base } : {}),
+        onBuildIdMismatch: () => {
+          invalidate();
+          manager.clear();
+        },
+      }),
+    options,
+  );
+};
 
 export const hasCachedShell = (
   route: RouteProps,
   currentElements: Record<string, unknown>,
-): boolean => singleton.hasCachedShell(route, currentElements);
+): boolean =>
+  canCommitInstantly(
+    getRouteSlotId(route.path),
+    currentElements,
+    getPrefetchedElements(route),
+  );
 
 export const getPrefetchedElements = (
   route: RouteProps,
-): Elements | undefined => singleton.getPrefetchedElements(route);
+): Elements | undefined => manager.getElements(encodeRoutePath(route.path));
 
 export const getPrefetch = (route: RouteProps): PrefetchHandle | undefined =>
-  singleton.getPrefetch(route);
+  manager.get(encodeRoutePath(route.path), route.query);
 
 export const canReuseStaticRoute = (
   route: RouteProps,
   currentElements: Elements,
-): boolean => singleton.canReuseStaticRoute(route, currentElements);
+): boolean =>
+  staticPathSet.has(route.path) &&
+  getRouteSlotId(route.path) in currentElements;
 
 export const learnStaticFromElements = (
   elements: Record<string, unknown>,
-): void => singleton.learnStaticFromElements(elements);
+): void => {
+  const route = getRouteFromElements(elements);
+  if (route && isStaticFromElements(elements)) {
+    staticPathSet.add(route.path);
+  }
+};
 
 export const clearCaches = (): void => {
-  singleton.clear();
-};
-
-const registeredLazySlices = new Set<string>();
-
-export const registerLazySlice = (id: string): void => {
-  registeredLazySlices.add(id);
-};
-
-export const forEachRegisteredLazySlice = (fn: (id: string) => void): void => {
-  registeredLazySlices.forEach(fn);
-};
-
-export const clearRegisteredLazySlices = (): void => {
-  registeredLazySlices.clear();
+  manager.clear();
+  staticPathSet.clear();
 };
