+++ packages/waku/tests/router-route-state-hooks.test.tsx
@@ -16,17 +16,17 @@ import {
 import * as minimalClient from '../src/minimal/client.js';
 import { INTERNAL_ServerRoot } from '../src/minimal/client.js';
 import * as caches from '../src/router/client-core-utils/caches.js';
+import { clearCaches } from '../src/router/client-core-utils/caches.js';
+import { useHmrRefetch } from '../src/router/client-core-utils/hmr.js';
 import {
-  clearCaches,
-  clearRegisteredLazySlices,
-  registerLazySlice,
-} from '../src/router/client-core-utils/caches.js';
-import {
-  useHmrRefetch,
   useInitialRoute,
   useInitialRscParams,
-} from '../src/router/client-core-utils/route-state-hooks.js';
+} from '../src/router/client-core-utils/initial-route.js';
 import * as slice from '../src/router/client-core-utils/slice.js';
+import {
+  clearRegisteredLazySlices,
+  registerLazySlice,
+} from '../src/router/client-core-utils/slice.js';
 import {
   ROUTE_ID,
   encodeRoutePath,
