+++ packages/waku/tests/router-state.test.ts
@@ -1,16 +1,13 @@
 /** @vitest-environment happy-dom */
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import { ETAG_ID_PREFIX, IMMUTABLE_ETAG } from '../src/lib/utils/etags.js';
 import {
   ROUTER_STATE_ID,
   getRouterState,
   getSettledRoute,
   makeRouterState,
-  pinForSwr,
   resolveServerRedirect,
 } from '../src/router/client-utils/router-state.js';
 import {
-  HAS404_ID,
   IS_STATIC_ID,
   ROUTE_ID,
 } from '../src/router/isomorphic-utils/route-path.js';
@@ -222,32 +219,5 @@ describe('getSettledRoute', () => {
     );
 
     expect(getSettledRoute(elements, fallback)).toEqual(settledRoute);
-  });
-});
-
-describe('pinForSwr', () => {
-  const immutable = (slotId: string) => ({
-    [ETAG_ID_PREFIX + slotId]: IMMUTABLE_ETAG,
-  });
-
-  test('pins meta keys and immutable slots, not mutable ones', () => {
-    const pin = pinForSwr(() => immutable('layout:/'));
-    expect(pin(ROUTE_ID)).toBe(true);
-    expect(pin(HAS404_ID)).toBe(true);
-    expect(pin(IS_STATIC_ID)).toBe(true);
-    // a legacy-style prefix is not meta; only the exact IS_STATIC key is
-    expect(pin(`${IS_STATIC_ID}:layout:/`)).toBe(false);
-    // the client's own state rides the merge instead of becoming a hole
-    expect(pin(ROUTER_STATE_ID)).toBe(true);
-    expect(pin('layout:/')).toBe(true);
-    expect(pin('page:/a')).toBe(false);
-  });
-
-  test('reads the resolved elements at call time', () => {
-    let resolved: Record<string, unknown> = {};
-    const pin = pinForSwr(() => resolved);
-    expect(pin('layout:/')).toBe(false);
-    resolved = immutable('layout:/');
-    expect(pin('layout:/')).toBe(true);
   });
 });
